+++ client/src/functions.ts
@@ -1,3 +1,5 @@
+import { isString, isNumber } from "util"
+
 export const pick = <T, K extends keyof T>(name: K) => (x: T): T[K] => x[name]
 export const flat = <T>(a: T[][]): T[] =>
   a.reduce((res, current) => [...res, ...current], [])
@@ -28,6 +30,25 @@ export const selectMap = <T1, K extends keyof T1, T2>(
   return ((record && record[property]) || defval) as T2
 }
 
+// tslint:disable-next-line: ban-types
+const isFn = (f: any): f is Function => {
+  return typeof f === "function"
+}
+
+export const mapGet = <T1, T2>(
+  map: Map<T1, T2>,
+  key: T1,
+  init: (() => T2) | T2
+): T2 => {
+  let result = map.get(key)
+  if (!result) {
+    result = isFn(init) ? init() : init
+    map.set(key, result)
+  }
+
+  return result
+}
+
 export const stringOrder = (s1: any, s2: any) => {
   if (s1 > s2) return 1
   return s2 > s1 ? -1 : 0
@@ -37,3 +58,18 @@ export const fieldOrder = <T>(fieldName: keyof T, inverse: boolean = false) => (
   a1: T,
   a2: T
 ) => stringOrder(a1[fieldName], a2[fieldName]) * (inverse ? -1 : 1)
+
+export function parts(whole: any, pattern: RegExp): string[] {
+  if (!isString(whole)) return []
+  const match = whole.match(pattern)
+  return match ? match.slice(1) : []
+}
+
+export function toInt(raw: any): number {
+  if (isNaN(raw)) return 0
+  if (isNumber(raw)) return Math.floor(raw)
+  if (!raw && !isString(raw)) return 0
+  const n = Number.parseInt(raw, 10)
+  if (isNaN(n)) return 0
+  return n
+}