+++ client/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to the "vscode-abap-remote-fs" extension will be documented 
 
 Format based on [Keep a Changelog](http://keepachangelog.com/)
 
+## Unreleased
+
+### Changed
+
+- delete favourites now displayed online
+- only show delete favourite for those who can be deleted
+- command descriptions - all starting with ABAPfs now
+
 ## [0.5.17] 2019-03-15
 
 ### Added