+++ client/CHANGELOG.md
@@ -6,6 +6,10 @@ Format based on [Keep a Changelog](http://keepachangelog.com/)
 
 ## Unreleased
 
+### Added
+
+- Abap Unit
+
 ### Changed
 
 - delete favourites now displayed online